Add abstract_component_test.py to fondant
