+++ components/caption_images/tests/test_caption_images.py
@@ -1,56 +1,7 @@
-
-from abc import ABC, abstractmethod
-
 import pandas as pd
-import pytest
 import requests
 from caption_images.src.main import CaptionImagesComponent
-
-
-class AbstractComponentTest(ABC):
-    @abstractmethod
-    def create_component(self):
-        """
-        This method should be implemented by concrete test classes
-        to create the specific component
-        that needs to be tested.
-        """
-        raise NotImplementedError
-
-    @abstractmethod
-    def create_input_data(self):
-        """This method should be implemented by concrete test classes
-        to create the specific input data.
-        """
-        raise NotImplementedError
-
-    @abstractmethod
-    def create_output_data(self):
-        """This method should be implemented by concrete test classes
-        to create the specific output data.
-        """
-        raise NotImplementedError
-
-    def setUp(self):
-        """
-        This method will be run before each test method.
-        Add any common setup steps for your components here.
-        """
-        self.component = self.create_component()
-        self.input_data = self.create_input_data()
-        self.expected_output_data = self.create_output_data()
-
-    def tearDown(self):
-        """
-        This method will be run after each test method.
-        Add any common cleanup steps for your components here.
-        """
-
-    def test_transform(self):
-        """
-        Default test for the transform method.
-        Tests if the transform method executes without errors.
-        """
+from fondant.abstract_component_test import AbstractComponentTest
 
 
 class TestCaptionImagesComponent(AbstractComponentTest):
@@ -74,18 +25,3 @@ class TestCaptionImagesComponent(AbstractComponentTest):
         return pd.DataFrame(
             data={("captions", "text"): {0: "a motorcycle", 1: "a beetle"}},
         )
-
-    @pytest.fixture(autouse=True)
-    def __setUp(self):
-        self.component = self.create_component()
-        self.input_data = self.create_input_data()
-        self.expected_output_data = self.create_output_data()
-
-    def test_transform(self):
-        output = self.component.transform(self.input_data)
-        assert output.equals(self.expected_output_data)
-
-    def tearDown(self):
-        del self.component
-        del self.input_data
-        del self.expected_output_data
